Consolidate indicator metadata helpers
